move client_name_by_epic_id function to epics data file and update paths
