+++ frontend/pages/data/epics.py
@@ -12,3 +12,13 @@ def epics_names() -> List[Select]:
         d = Select(value=item["id"], display_value=item["name"])
         epic_name_rows.append(d)
     return epic_name_rows
+
+
+def client_name_by_epic_id(epic_id) -> Select:
+    api_client_name_id = f"{base_url}/api/epics/{epic_id}/client-name"
+    response_client_name_id = requests.get(api_client_name_id)
+    r = response_client_name_id.json()
+    client_name = r.get("name")
+    client_id = r.get("id_1")
+    d = Select(value=client_id, display_value=client_name)
+    return d
